backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.36.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.9
python-dotenv==1.1.1
python-multipart==0.0.9

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-cov==5.0.0

